+++ responses/matchers.pyi
@@ -1,0 +1,14 @@
+from typing import (
+    Any,
+    Callable,
+    Optional,
+    Dict,
+)
+
+def json_params_matcher(
+    params: Optional[Dict[str, Any]]
+) -> Callable[..., Any]: ...
+
+def urlencoded_params_matcher(
+    params: Optional[Dict[str, str]]
+) -> Callable[..., Any]: ...
